finished ex 2.15*

+++ part2/phonebook/src/App.js
@@ -33,7 +33,7 @@ const Filter = ( {searchQuery, handleSetSearchQuery} ) => {
 
 const PersonForm = (props) => {
   return(
-    <form onSubmit={props.addPerson}>
+    <form onSubmit={props.handleAdd}>
       <div> name: <input value={props.newName} onChange={props.handleNameChange} /> </div>
       <div> number: <input value={props.newNumber} onChange={props.handleNumberChange} /> </div>
       <div> <button type="submit">add</button> </div>
@@ -57,22 +57,25 @@ const App = () => {
       .catch(error => alert("failed getting initial persons"))
     }, [])
   //-----event handlers-----------------------------------------
-  const addPerson = (event) => {
-    event.preventDefault()
-    //console.log("event.target: ",event.target, " event.target.value: ",event.target.value)
-    /*const names = persons.map(v => v.name) //from array of object transform into array of contents of 'name' property of the object
-    if(names.includes(newName)) {
-      alert(`${newName} is already added to phonebook`) //alert(newName, "is already added to phonebook") does not work becasue console.log() just can accept more arguments
-      return;
-    }*/
-    if(persons.some( person => person.name === newName )) {
-      alert(`${newName} is already added to phonebook`) //alert(newName, "is already added to phonebook") does not work becasue console.log() just can accept more arguments
-      return;
-    }
+
+  const changeNumber = (existingPerson) => {
+    if(window.confirm(`Do you want to change the person ${existingPerson.name} number from ${existingPerson.number} to ${newNumber}?`)) {
+      const changedPerson = {...existingPerson, number: newNumber}
+      
+      personsService
+        .update(existingPerson.id,changedPerson) //update server
+        .then(returnedPerson => {
+          setPersons(persons.map(prsn => prsn.id !== existingPerson.id ? prsn : returnedPerson))
+        }).catch(error => alert("failed updating person"))
+    } 
+  }
+
+  const addPerson = () => {
     const newPerson = {
       name: newName,
       number: newNumber
     }
+
     personsService
       .create(newPerson)
       .then(returnedPerson => {
@@ -81,7 +84,28 @@ const App = () => {
         setNewNumber('')
       })
       .catch(error => alert("failed adding person"))
+  }
 
+  const handleAdd = (event) => {
+    event.preventDefault()
+    /*
+    //console.log("event.target: ",event.target, " event.target.value: ",event.target.value)
+    const names = persons.map(v => v.name) //from array of object transform into array of contents of 'name' property of the object
+    if(names.includes(newName)) {
+      alert(`${newName} is already added to phonebook`) //alert(newName, "is already added to phonebook") does not work becasue console.log() just can accept more arguments
+      return;
+    }*/
+
+    const existingPerson = persons.find( person => person.name === newName )
+    if(existingPerson) {
+      if(existingPerson.number === newNumber) {
+        alert(`${existingPerson.name} with number ${existingPerson.number} is already added to phonebook. You can swap the number or add another person.`) 
+        //alert(newName, "is already added to phonebook") does not work becasue console.log() just can accept more arguments
+        return;
+      } 
+      else { changeNumber(existingPerson) }
+    }
+    else { addPerson() }
   }
 
   const handleNameChange = (event) => {
@@ -100,14 +124,18 @@ const App = () => {
 
   const deleteEntry = id => {
     //console.log(`delete ${id}`)
-    personsService
-      .deletePerson(id)
-      .then(() => setPersons(persons.filter(person => person.id !== id))
-      )
-      .catch(error => {
-        alert(`could not process deletion of person with id ${id}`)
-        //no action here because we don't yet know if that person was not in the database, or it was but some other (?) error occured - so we dont know whether we should render that person client-side or not - thus we take no action.
-      })
+    const personName = persons.find(prsn => prsn.id === id).name
+    if(window.confirm(`Do you really want to delete person ${personName}`))
+    {
+      personsService
+        .deletePerson(id)
+        .then(() => setPersons(persons.filter(person => person.id !== id))
+        )
+        .catch(error => {
+          alert(`could not process deletion of person ${personName}`)
+          //no action here because we don't yet know if that person was not in the database, or it was but some other (?) error occured - so we dont know whether we should render that person client-side or not - thus we take no action.
+        })
+    }
   }
   //------------------------------------------------------------
   
@@ -120,7 +148,7 @@ const App = () => {
       
 
       <h3>Add new entry</h3>
-      <PersonForm addPerson={addPerson} newName={newName} handleNameChange={handleNameChange} newNumber={newNumber} handleNumberChange={handleNumberChange} />
+      <PersonForm handleAdd={handleAdd} newName={newName} handleNameChange={handleNameChange} newNumber={newNumber} handleNumberChange={handleNumberChange} />
 
       <h3>Numbers</h3>
       <Persons persons={persons} searchQuery={searchQuery} deleteEntry={deleteEntry} />
